Split BUILD files
