+++ testsuite/bzltestsuite/java/jflex/testing/BUILD
@@ -1,0 +1,13 @@
+package(default_visibility = ["//testsuite:__subpackages__"])
+
+java_library(
+    name = "testing",
+    srcs = glob(["*.java"]),
+    deps = [
+        "//jflex",
+        "//testsuite/bzltestsuite/java/jflex/testing/annotations",
+        "//testsuite/bzltestsuite/java/jflex/testing/javac",
+        "//third_party/com/google/guava",
+        "//third_party/junit",
+    ],
+)
