+++ testsuite/bzltestsuite/java/jflex/testing/JFlexTestRunner.java
@@ -6,7 +6,8 @@ import com.google.common.collect.ImmutableList;
 import java.io.File;
 import jflex.LexGenerator;
 import jflex.testing.annotations.TestSpec;
-import jflex.testing.exception.CompilerException;
+import jflex.testing.javac.CompilerException;
+import jflex.testing.javac.JavacUtil;
 import org.junit.runner.Description;
 import org.junit.runner.notification.Failure;
 import org.junit.runner.notification.RunNotifier;
