+++ testsuite/bzltestsuite/java/jflex/testing/annotations/BUILD
@@ -1,0 +1,6 @@
+package(default_visibility = ["//testsuite:__subpackages__"])
+
+java_library(
+    name = "annotations",
+    srcs = glob(["*.java"]),
+)
