+++ testsuite/bzltestsuite/java/jflex/testing/javac/BUILD
@@ -1,0 +1,9 @@
+package(default_visibility = ["//testsuite:__subpackages__"])
+
+java_library(
+    name = "javac",
+    srcs = glob(["*.java"]),
+    deps = [
+                "//third_party/com/google/guava",
+    ]
+)
