+++ testsuite/javatests/jflex/testcase/apipirivate/BUILD
@@ -1,9 +1,14 @@
 java_test(
     name = "ApiPrivateTest",
     srcs = ["ApiPrivateTest.java"],
-    data = ["private.flex", "AttemptPrivateAccess.java"],
+    data = [
+        "AttemptPrivateAccess.java",
+        "private.flex",
+    ],
     deps = [
         "//testsuite/bzltestsuite",
+        "//testsuite/bzltestsuite/java/jflex/testing/annotations",
+        "//testsuite/bzltestsuite/java/jflex/testing/javac",
         "//third_party/com/google/guava",
         "//third_party/com/google/truth",
     ],
